Combined MapActivity with MapFiltersActivity
