--- app/src/main/java/edu/wit/mobileapp/hiddengem/MapActivity.java
@@ -1,42 +1,0 @@
-package edu.wit.mobileapp.hiddengem;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import com.google.android.gms.location.places.Place;
-import com.google.android.gms.maps.CameraUpdate;
-import com.google.android.gms.maps.CameraUpdateFactory;
-import com.google.android.gms.maps.GoogleMap;
-import com.google.android.gms.maps.OnMapReadyCallback;
-import com.google.android.gms.maps.SupportMapFragment;
-import com.google.android.gms.maps.model.LatLng;
-import com.google.android.gms.maps.model.MarkerOptions;
-
-public class MapActivity extends AppCompatActivity implements OnMapReadyCallback {
-    LatLng userPlaceLatLng;
-    String userPlaceAddress;
-    String userPlaceName;
-
-    @Override
-    protected void onCreate(final Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_map);
-
-        final Bundle previousBundle = getIntent().getExtras();
-        final double latitude = previousBundle.getDouble("place_latitude");
-        final double longitude = previousBundle.getDouble("place_longitude");
-        userPlaceLatLng = new LatLng(latitude, longitude);
-        userPlaceAddress = previousBundle.getString("place_address");
-        userPlaceName = previousBundle.getString("place_name");
-
-        final SupportMapFragment mapFragment = (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map);
-        mapFragment.getMapAsync(this);
-    }
-
-    @Override
-    public void onMapReady(final GoogleMap googleMap) {
-        googleMap.addMarker(new MarkerOptions().position(userPlaceLatLng).title(userPlaceName));
-        googleMap.moveCamera(CameraUpdateFactory.newLatLng(userPlaceLatLng));
-        googleMap.moveCamera(CameraUpdateFactory.zoomTo(googleMap.getMaxZoomLevel() * 0.6f));
-    }
-}
